Create a notebook predicting COVID-19 testing
This attempts to predict the number of COVID-19 tests using ergo, while
trying to demonstrate how you can use modeling with judgement to
iteratively improve a model. I tried to make it a gentle starting point
for new folks to `ergo`.

It could be more sophisticated, but opted for simplicity here since I
wanted to try mimic how a lazy-ish forecaster would actually use it.